Policy Implementation

+++ app/Http/Controllers/CategoryController.php
@@ -9,6 +9,8 @@ use App\Http\Traits\RespondsTrait;
 class CategoryController extends Controller
 {
     use RespondsTrait;
+
+
     public function index()
     {
         $category = Category::get();
@@ -17,36 +19,36 @@ class CategoryController extends Controller
 
     public function store(CategoryRequest $request)
     {
+
         $data['name'] = $request->name;
+        $data['admin_id'] = auth()->user()->id;
         Category::create($data);
         return $this->successResponse('Category Added Successfully', $data, 200);
     }
 
-    public function show($id)
+    public function show(Category $category)
     {
-
-        $category = Category::findOrFail($id);
-        return $this->successResponse('This is the Category with id = ' . $id, $category, 200);
+        $this->authorize('view', $category);
+        return $this->successResponse('This is the Category with id = ' . $category->id, $category, 200);
     }
 
 
-    public function update(CategoryRequest $request, $id)
+    public function update(CategoryRequest $request, Category $category)
     {
-
-        $category = Category::whereId($id);
+        $this->authorize('update', $category);
         $data['name'] = $request->name;
         $category->update($data);
 
-        return $this->successResponse('Category with id = ' . $id . ' Updated Successfully', $data, 200);
+        return $this->successResponse('Category with id = ' . $category->id . ' Updated Successfully', $data, 200);
     }
 
-    public function destroy($id)
+    public function destroy(Category $category)
     {
-        $category = Category::findOrFail($id);
-        if($category->product){
-            return $this->errorResponse('Category with id = ' . $id . ' cannot be deleted due to dependency',422 );
+        $this->authorize('delete', $category);
+        if ($category->product) {
+            return $this->errorResponse('Category with id = ' . $category->id  . ' cannot be deleted due to dependency', 422);
         }
         $category->delete();
-        return $this->successResponse('Category with id = ' . $id . ' deleted Successfully', $category, 200);
+        return $this->successResponse('Category with id = ' . $category->id . ' deleted Successfully', $category, 200);
     }
 }
